backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
google-generativeai==0.8.5
httpx==0.28.1
